+++ js/arrays.js
@@ -1,0 +1,15 @@
+const ESTATE_TYPE = [
+  'palace', 'flat', 'house', 'bungalow', 'hotel',
+];
+const TIME_TO_CHECK = [
+  '12:00', '13:00', '14:00',
+];
+const SAMPLE_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const SAMPLE_FEATURES = [
+  'wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner',
+];
+export {ESTATE_TYPE,TIME_TO_CHECK,SAMPLE_PHOTOS,SAMPLE_FEATURES};
